16-bit hidden layer weights
Bench: 18135035

+++ src/nnue/network.rs
@@ -95,7 +95,7 @@ pub static NNUE: NNUEParams =
 pub struct NNUEParams {
     pub feature_weights: Align64<[i16; INPUT * LAYER_1_SIZE * BUCKETS]>,
     pub feature_bias: Align64<[i16; LAYER_1_SIZE]>,
-    pub output_weights: Align64<[i8; LAYER_1_SIZE * 2]>,
+    pub output_weights: Align64<[i16; LAYER_1_SIZE * 2]>,
     pub output_bias: i16,
 }
 
@@ -590,7 +590,7 @@ impl<const MIN: i16, const MAX: i16> ActivationFunction for SquaredClippedRelu<M
 fn flatten<F: ActivationFunction>(
     us: &Align64<[i16; LAYER_1_SIZE]>,
     them: &Align64<[i16; LAYER_1_SIZE]>,
-    weights: &Align64<[i8; LAYER_1_SIZE * 2]>,
+    weights: &Align64<[i16; LAYER_1_SIZE * 2]>,
 ) -> i32 {
     let mut sum: i32 = 0;
     for (&i, &w) in us.iter().zip(&weights[..LAYER_1_SIZE]) {
